Functionality Updates
1. Animation added between display “none” & “block” on the student
items. This was done by adding and removing an opacity class while
simultaneously utilizing setTimeouts to ensure that the display
variable fires first or second depending on the desired outcome.
2. Pagination links have been positioned at the bottom of the .page
element so that they don’t move once the student elements are removed
from the DOM.
3. A fix was added around what the margin-left would be for the
Pagination links parent element ‘.pagination’. This keeps the element
centered when they disappear and reappear.
4. Condition added within the createPageLinks function which removes
all links if the parameter passed equals 1.