Adjust Pagination ReWrite
I rewrote the pagination builder so that the elements will build the
maximum required once and only a class add and or remove will display
or hide the elements moving forward. This should reduce the work load
required when running this particular function. I’ve also simplified a
lot of my notes as they seemed excessive as well as cleaned up any
general inefficiencies lying around.